+++ Runtime/LensPackageInfo.cs
@@ -1,0 +1,8 @@
+namespace KostasBan.Lens
+{
+    public static class LensPackageInfo
+    {
+        public const string PackageName = "com.kostasban.lens";
+        public const string Version = "0.4.0";
+    }
+}